Module Database is updated,  which implement saving user data in json file
